+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
   * Fix a bug that could cause `WEAK` symbols to be not exported
   * Skip binary printing interpreters when printing multiple multiple modules
+  * Require gtirb >=2.1.1
 
 # 2.2.0