+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
   * Fix a bug that could cause `WEAK` symbols to be not exported
   * Skip binary printing interpreters when printing multiple multiple modules
-  * Require gtirb >=2.1.1
+  * Require gtirb >=2.2.0
 
 # 2.2.0